requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn
scipy>=1.7.0
emcee
PyYAML>=6.0
corner
h5py>=3.0.0